+++ operation/src/main/scala/operation/lifting/GroupingParsers.scala
@@ -1,0 +1,29 @@
+package com.faacets
+package operation
+package lifting
+
+import data._
+import core._
+import fastparse.noApi._
+import net.alasc.domains.Partition
+
+object GroupingParsers {
+
+  import com.faacets.data.Parsers._
+
+  import White._
+
+  val inputNoGrouping: P[InputGrouping] = int.map( InputGrouping.noLifting(_) )
+
+  val inputWithGrouping: P[InputGrouping] = P( "{" ~ int.rep(min=1) ~ "}" ).map { outputs =>
+      val blocks: Iterable[Set[Int]] = outputs.zipWithIndex.groupBy(_._1).map { case (k, v) => v.map(_._2).toSet }
+      InputGrouping(Partition(blocks.toSeq: _*))
+  }
+
+  val inputGrouping: P[InputGrouping] = P( inputNoGrouping | inputWithGrouping )
+
+  val partyGrouping: P[PartyGrouping] = P( "(" ~ inputGrouping.rep(min=1) ~ ")" ).map(PartyGrouping(_))
+
+  val grouping: P[Grouping] = P( "[" ~ partyGrouping.rep(min=1) ~ "]").map(Grouping(_))
+
+}
